Refactor: merge ghost logic into single scary_ghost function
Duplicate code for the most part in milky potion and haunted temple
ghosts. This also changes the logic of a milky ghost to be consistent
with the temple one: if you can spot the ghost, you get scared; being
blind and telepathic/monster-detecting doesn't save you.

This commit should have come earlier in the history than "Ghost
overhaul", but a bad rebase apparently removed it, so the history here
is messed up.