Most things that affected shades getting hit now apply to ghosts too
Strictly speaking, they now apply to all incorporeal monsters, but only
ghosts and shades are. This is partially for flavor (ghosts' method of
being hard to touch with material implements was just... having very low
AC, which meant a late-game character could take a perfectly mundane
weapon and hit them with it every time), and partially as a first step
to changing overall ghost behavior. More on this in the next commits.

In the process of this I discovered that monster weapon attacks and
monster and player unarmed attacks both hurt shades, which carries over
to ghosts, but that seems weird to me. I've asked the devteam about it
and deliberately didn't change anything regarding that behavior here,
in case they decide to address it.