Ghost overhaul
Instead of coming up to you and making ineffectual 1d1 touch attacks ad
infinitum, they now turn invisible when able to and follow the player,
trying to get adjacent and turn visible, basically jump-scaring the
player. If they succeed, the player is frightened and paralyzed the same
as a ghost from a bottle or temple. The player can prevent this by
having some method of seeing the ghost such as see invisible or detect
monsters; even blinding themselves will prevent the sudden scare.

Internally this is done by giving them a "touch paralyze" attack that
isn't strictly a touch or a (direct) paralysis, but it's the closest
counterpart.

The core statblock for ghosts is also changed, making them less tanky
but also faster. See the changelog for full details.

The ultimate reason for this change is in the hopes of making ghosts a
more distinct monster than the boring behavior they used to exhibit.