Sharply reduce floating eye paralysis severity
(level+1)d70 is way too high for a balanced monster of difficulty 3.
Reduce this to (level+1)d10, and it always freezes for that amount, not
randomly deciding to freeze for 127 turns. Also, add in a Wisdom saving
throw of sorts - roll d30, freeze only if the roll is higher than
Wisdom. Otherwise, you avert your eyes and nothing happens.

There's possibly room for implementing a later-game e monster that has
longer, active paralysis gaze.