Levels beyond 45 don't add to score
This is to make minscore players happy by making a minimum score
ascension achievable in any game, not just a game where the Sanctum
happens to be on level 45. I'm not aware of anyone who would be angry
that you now get constant points for the depth of the Sanctum instead of
more for a deeper one.