Revert water elemental/gelcube suffocation
This reverts commits ee808b7, 6602893, and bbf7557.

Players brought it to my attention that this can immediately instakill
the player two ways: gel cube paralyzes you for 8 turns, engulfs you,
and you die in 6, and water elementals engulf you and drag you out over
deep water on the Plane of Water; when you kill them you drown (this has
nothing to do with the suffocation; even if they only did water damaging
engulfs this would be a hazard, since the player can't verify whether or
not they've been moved over water). It was also pointed out that the
beefed up Plane of Water elementals would be hard to kill conventionally
before you suffocate, and hard to avoid, especially with the bubbles
moving around.

As such, while it did make water elementals more threatening and
gelatinous cubes more flavorful, the current incarnation of this feature
isn't properly balanced and needs more time to be looked at.