Chameleons can no longer turn into out-of-depth monsters
They still have a 1/3 chance to pick an animal form, but they won't pick
an animal form that is out of depth either. The remainder of the time,
if any shapechanging monster fails to choose a specific form, it will
pick a random monster as before - but the random monster must be
in-depth.

The main reason for this is because no one likes chameleon arch-liches
(or mumaks, or titans, or balrogs, etc...) appearing out of nowhere in
the early game to kill them.