Refactor: rndmonst() simplification
Using weighted reservoir sampling rather than a unwieldy pre-computed
cache of eligible monsters that needs to be updated every time the hero
goes up or down a floor, gains or loses a level, or extincts or
genocides a monster.

This still requires only a single pass over the list of monsters (the
same as what rndmonst() required even if it didn't need to update the
cache), and is able to get rid of all the cache code.