Give player message informing them they can use #enhance
Triggers when you feel more confident in your skills. This is to address
a problem I have heard about several times from newer players: unless
you pay close attention to the guidebook, nothing in the game actually
indicates that you can level up your abilities and how to do it.

Eventually, if I keep adding helpful hints for new players like this,
there should probably be some option that disables them (this and
similar messages are just clutter for an experienced player).
I don't think this is common enough to seriously irk anyone at the
moment, but if it does, at least this can be hidden via MSGTYPE.