Revert extra fuzzing of monster group size generation
This reverts part of bd9194d, the part that changed the constants in
m_initsgrp and m_initlgrp. I don't think I realized at the time that it
was already randomized.

Return group size generation to vanilla rules. This doesn't touch the
other parts of that commit in which create_critters forces even group
monsters to generate individually.