Blessed and uncursed wands wrest much more often
Like in UnNetHack, blessed wands wrest 1/7 of the time, and uncursed
wrest 1/23 of the time. Cursed wands still wrest 1/121 of the time. This
is intended as a minor buff to having a blessed wand, and making it
feasible to wrest uncursed wands without spending fifty or more turns
zapping it to no effect.

Ideally, I would have blessed wands identify that they have 0 charges
when you zap them and nothing happens (and never wrest if you didn't
know it had 0 charges), to prevent an accidental 1/7-probability wrest.
However, there's no way to have the charges be known without also having
the recharge count be known, which doesn't make any sense for the hero
to know. And it's certainly not worth putting in another known flag just
for this.

The only place where I can see this biting someone is that if you tend
to zap a wand of wishing until you get "Nothing happens" without
actually identifying its charge count, there's now a higher chance of
you wresting it. However, the overall risk is the same; following this
approach under the old behavior could still randomly wrest it. Plus,
planned changes to the wand of wishing will make wresting them a
non-issue in the future, so it's not that big of a deal.