Scatter wishes throughout the late game
This is a very strategically significant and soon to be controversial
change. It is intended to address one of the key components of why the
late game is commonly held to be a boring slog, viz. that the player
gets a large dump of 5 to 7 wishes at the castle that they can then use
to wish up all the equipment they need to ascend, so they can coast
through Gehennom with all the gaps in their defenses covered.

To address this, the player now gets only two wishes at the castle, and
will need to delve into Gehennom to get more. This is done by making
some changes to wands of wishing: they turn to dust after they spend
their last charge (making them unwrestable) and they explode whenever
you try to recharge them. Then, any wand with X charges will give
exactly X more wishes, without variance.

The castle wand now has 2 charges: wands with one charge are placed in
Vlad's chest, with the Wizard, in the Fake Wizard's Tower, and in one of
the rooms in the Sanctum. Wands of wishing no longer randomly generate.