Show the weights of objects in inventory
A quality-of-life feature that had been requested for a while, but which
I didn't know how to go about implementing until bhaak pointed out the
relevant bits in the UnNetHack code for this. Display of the weights is
now controlled by the "invweight" option, defaulting to true and usable
outside wizard mode.

This implementation differs from UnNetHack's in a few ways:
  1. It only works on items in inventory, unless in wizard mode in which
     case all items' weights are seen. This prevents the player from
     detecting loadstones on the ground.
  2. It doesn't have special cases that exclude gray stones, rocks,
     statues, boulders, and other post-rock objects. I didn't see a need
     for these; if the player picks up a statue or heavy iron ball (or a
     loadstone), let them see its weight.
  3. Weights for box-type containers seemed to be fine with my testing
     (and Schroedinger's Box should be fine too), so I didn't put in
     unnethack's special exclusions for those, either.
  4. Hallucination doesn't block weight display like it does in
     UnNetHack.

This was subject to a display bug from UnNetHack in which an item that
gets added to the inventory and merged with other objects will show the
total merged weight, not the weight of the items just obtained, but
bhaak provided a fix which is included in this commit.