Naming artifacts no longer breaks illiterate conduct
There were several flavor problems with the previous behavior of this.
It insinuated that naming an item entails engraving the user-provided
string onto the item. This can be done even if the hero has nothing else
to engrave with; it can be done without hands to do the engraving
(handless polyforms can't engrave onto the floor); it can be done in
zero turns and an infinite amount of times regardless of the string
length; it can be done while blind or hallucinating with perfect
accuracy; and so on.

Clearly, minus the artifact case, the naming of items is just a helpful
interface abstraction, and attempts to make it work in context of the
hero actually doing something are doomed to lots of problematic flavor
like this.

It's also worth noting that Sting - one of only two nameable artifacts
in the game - was pretty obviously just named by Bilbo deciding to call
it that, and not writing on it somehow.

Then there's a whole inconsistency in that naming the item "Stinf", or
any other non-artifact name, somehow doesn't break illiterate conduct -
but trying to give it an artifact name, whether or not this fails, will
break it. There is a comment in the source that says maybe illiterate
should be violated for naming items anything other than X, but then this
just turns into an interface screw for illiterate-conduct players, by
removing this helpful abstraction.

So to clear all this up, the new flavor of naming items is that the hero
simply mentally assigns a name to an item and remembers it. Therefore,
illiterate conduct is not violated by trying to name an artifact.
Attempting and failing now produces the message "For some reason, you
can't mentally assign that name to it."

There are still a couple minor flavor issues with this approach - one is
that if, say, a dagger falls down a hole, how would you recognize it
versus a new dagger? But the game already has this problem with e.g. if
the dagger's enchantment were known. Another issue is how the player
knows the name of a randomly generated named weapon, but again this had
problems before; even if the "Spear of Doom" had that written on it, the
player's getting the message with that name while watching a hobbit from
all the way across the Big Room wield it. In sum: this approach fixes
flavor issues while not creating any new ones.