Incorporate Malcolm Ryan's Brewing Patch
This was an old patch versus 3.4.3, but was not hard to port to 3.6. The
main addition of the patch enables the player to dip mold corpses in
potions of fruit juice to start a fermentation timer that will
eventually turn the potion into something else dependent on the mold.

Since xNetHack makes molds pretty easy to find due to their growing on
corpses, green mold does not produce healing as it does in the original
patch; it produces acid. Additionally, the booze and sleeping potions
are swapped; booze is now produced from red mold and sleeping from brown
mold. As in the original patch, lichens and shriekers and non-F corpses
do not ferment. It also doesn't show when a fermenting potion changes in
a monster's inventory.

This also makes the shopkeeper charge you for shop-owned fruit juice
that you begin fermenting. The original patch didn't have that, but I
tested without it and it allows you to buy the new, possibly more
expensive potion at the cost of the fruit juice. To simplify things,
just make it charge at the start of fermentation rather than the end.