Major buff to Ogresmasher
The most significant thing is that it is now an ogre-bane in name only;
it no longer specifically targets ogres. I may end up changing the name
later, or not. It was languishing because its bonuses applied only to a
not very common nor threatening class of monsters; making it not an
actual bane may help with this.

Original idea came from one of the users on discord, way back in 2017.
The idea was that the best buff for Ogresmasher is simply to make it an
outsized sledgehammer that splatters everything it hits and disables
larger monsters. To this end, on top of the new increased war hammer
damage, it gets +d3 to hit and +d8 versus all monsters, and additionally
has "heavy critical hits". Like the Tsurugi and Vorpal Blade, this is on
1/20 of hits, and a critical from Ogresmasher will instakill a small or
medium monster, and stun and slow down a medium or larger monster.

Its passive bonus of conferring 25 Constitution while wielded is not
changed.