Change the base types of Dragonbane and Demonbane
Dragonbane is now a dwarvish spear (because dwarves probably fight
dragons often enough to have made a spear for it), but mainly to get
away from most artifact weapons being swords of some sort.

Demonbane is now a mace instead of a long sword, for similar reasons,
and because this makes it particularly apropos for Priests, especially
before the Quest.