Make the Sceptre of Might the Priest quest artifact
Briefly:
1. The Mitre of Holiness is removed.
2. The Priest quest artifact is now the Sceptre of Might.
3. The Caveman quest artifact is now Big Stick, from NetHack Fourk.

The main impetus here is to make the choice of Priest alignments more
interesting because the Sceptre deals double damage to cross-aligned
monsters. As a Caveman artifact, it was just a reason not to pick
neutral if you could help it. I don't have any particular problems with
the Mitre, but it's ultimately rather bland strategically speaking.

The new Sceptre retains the Mitre's invoke power of an energy boost, but
now it confers drain resistance when wielded instead of magic
resistance.

Big Stick, the new Caveman quest artifact, is taken from Fourk pretty
unmodified. It has +d5 to-hit and +d12 to all monsters, and confers
stealth when carried and magic resistance when wielded. As the Sceptre
used to be able to do, the invoke power toggles hungerless conflict.
Unusually, it is a chaotic artifact despite Cavemen not being startable
as chaotic.

"The Sceptre of Might" doesn't really sound like the best name for the
Priest quest artifact anymore; it should sound holier. Not picking a new
name for it right now because 1) I can't think of anything good sounding
and 2) I'm wary of annoying people who want to wish for it.

Since it changes the artifacts in the artifact list, this is a
savebreaking change and will be one of the last changes in 0.4.0 before
0.5.0 starts development.