xNetHack version 0.4.1
In terms of code, this version doesn't exist. It's just a formality to
synchronize the version tracked by the code and the version tracked by
public servers (prior to this, servers called in-progress 0.4.0 as 0.4.0
as if it were released, but development of 0.4.0 was still ongoing).
Making an empty patchlevel version bump allows both to get on the same
page; the server can refer to the incomplete next version as 0.4.1 since
it is the last released version, and the patchlevel also tracks the
current version as 0.4.1.