Call potion bottles by nonsensical names if hallucinating
Recently a hitherto unknown fork of SpliceHack called 'SliceHack' was
discovered, and it contains some good stuff that xNetHack will be
incorporating. This is the first of many such features.

Note that this refers to the description of the physical bottle; it's a
substitute for "phial", "carafe", "flask", etc. such as are seen when a
potion crashes on someone's head. They don't obscure the randomized
appearance or actual potion identity.

The SliceHack version evidently went through several revisions; just
take the current one to avoid having multiple commits for this
relatively minor feature.