You may survive food poisoning with a Con/100 chance
Another SliceHack feature. At first I wasn't going to take this one, but
then I realized that the only time you get food poisoning is if you're
a new player who hasn't learned about tainted corpses yet or if you just
did something stupid. So, be a little nicer in those scenarios.

Surviving silently drops your Con by 1. But unlike in the original
commit, the chance is not (Con+Luck)/100, it's just Con/100, because I
don't want to add more effects that luck is a factor for.