New paranoid confirmation for throwing ammo without launcher
Another SliceHack feature. Trying to throw things like arrows, rocks,
and bolts without wielding the proper launcher for them will prompt for
a confirmation, since this attack will be ineffectual without using the
launcher and the player is probably making some kind of mistake.

Some people have previously imagined another solution to this that
involves some message like "You ineffectively throw one arrow". The
problem is that it's still a one-time interface screw, and since there's
a less annoying way to circumvent it, do that instead.