Hungerless casting ignores "too hungry to cast" penalty
Another SliceHack feature. This includes a refactor of one of the
leading contenders for "ugliest section of code in NetHack", the switch
statement determining spell hunger reduction. The refactor pulls the
hunger reduction logic into its own function, and provides a macro that
is used to check, essentially, whether you have hungerless casting.