A player killed by a zombie will arise from the grave as one
I was surprised to find that this feature isn't actually in vanilla. It
consists of the same undead-arising-from-grave logic as all the other
undead cases, except with zombies.

This being xNetHack, you have a chance of keeping your sanity in this
case (25% chance, higher than for any other undead).

Bonus: This commit makes use of the previously unused vanilla field
'urace.zombienum'.