Being Weak doesn't count as starving if you have enough food
Another SliceHack feature, but heavily modified. This changes the old
check for Weak hunger status into a more complicated starving() function
that factors in food the hero is carrying and whether they have slow
digestion. The aim of this is so that if you are Weak and happen to have
a monster critically injure you, your prayer usually won't go towards
fixing your hunger instead.

The original commit had a ton of complex logic for determining if the
food you're carrying is a valid candidate for eating, with
disqualifications for cockatrice corpses, globs of slime,
cannibalism-prone corpses, etc. This implementation doesn't do any of
that; it just blanket disqualifies all corpses, eggs, and globs. It also
doesn't disqualify low-nutrition foods like garlic. All it cares about
is the total sum of nutrition points you have.

The general rule is "if you don't have enough non-corpse non-egg
non-glob non-tin nutrition on your person to get you to NOT_HUNGRY
status, treat this as a major problem." It's harsher if the player has
slow digestion since that effectively gives them a much larger time
buffer. Another code bit not copied over from SliceHack is the one that
duplicates a ton of gethungry() code to calculate the player's effective
hungering rate right now; this algorithm doesn't care about that.

If the player is *trying* to get free food from their god, it's still
pretty easy to do that by just dropping all food and containers with
food in them.