Refactor: obj->material assignments into set_material()
Refactor taken from SliceHack. Simply setting the material didn't do
weight recalculation or handle erosion/erosionproofing differences
between the old and new material.

Setting material to the base material for that ought to be okay to leave
alone, but I turned all of them into set_material calls for
completeness. It might be called for to add a set_default_material() or
similar function, so the caller doesn't need to write out
"objects[obj->otyp].oc_material" all the time.

Since it's now a function, I added a sanity check that will impossible
if anything ever tries to set an object to a normally invalid material.
(But that'll have to change if we ever *want* to have items in the game
made out of normally invalid materials...)