Potions of hallucination can give enlightenment
Another SliceHack feature. Wasn't really inclined to take it, but
there's no gameplay downside to it and it gives a much-needed legitimate
use to one of the potions that's commonly blanked.

Oddly, uncursed hallucination potions will never give enlightenment; I'm
not sure why this is, but I kept the behavior.

Given moldy corpses, this will probably make enlightenment more common
in the early game for those who know how to brew it, which is also
probably a good thing since it gives more avenues for things like amulet
and ring identification.