Remove 'deferred' code that lets non-wizard player wish for materials
This feature was deferred "until we see the balance implications of obj
materials". It's been almost a year, and we have seen the balance
implications: they're fine as is.

A system where players can't choose to wish for materials has
significant advantages, primarily in encouraging the use of found
objects. If you are able to wish for materials, the game gets more
boring since in all likelihood you're going to end up using gear of an
optimal material that you've wished for. (See also: GruntHack.)

Insofar as the 3.6.x Object Materials Patch is concerned, the player
should not be able to specify the material of a wished object.