Grimtooth is now permanently poisoned
A buff present in several variants and Nephi's Bane Boosting patch, but
via SliceHack and its implementation. It is currently the only melee
weapon that can be poisoned.

There are significant implementation differences.

First, I noticed that SliceHack (and maybe the original patch?) didn't
provide for mhitu or mhitm combat with poisoned melee weapons
whatsoever, so I added those, behaving the same way as uhitm currently
does for poisoned weapons. Since xnethack got rid of the poison
instadeath for players and monsters both a while back, the current
behavior was to deal d6 damage to the player for thrown weapon attacks,
and d10+6 damage to the player otherwise. However, d10+6 is way too much
to be dishing out with each hit from Grimtooth, so I have added logic
that makes the poison from a poisoned melee weapon only take 25% of the
time.

Second, SliceHack (and maybe the original patch?) assumes that
otmp->oartifact suffices for a weapon being permanently poisoned as far
as if statements are concerned. Since it's possible that other
permanently poisoned artifacts or even regular objects could be added in
the future, I didn't want to do that. So this adds a one-line function
permapoisoned() which currently just checks if the object it was passed
is Grimtooth, but is otherwise easily extended.

Permanently poisoned objects are immune to losing their poison from
being dipped in a healing potion.