Deliberate level teleporter activation ignores magic resistance
Another SliceHack feature, ultimately from "Stevie-O's level teleporter
jump patch". I simplified it a bit: deliberately jumping onto the trap
always takes time even if it fails to levelport you (which would only
happen with level teleporters in the End Game, which don't exist).

This is aimed at providing a little quality of life in the form of not
having to divest yourself of your sources of magic resistance before
using a level teleporter.