Confused scrolls of earth summon earthy monsters
Another SliceHack feature, but again heavily modified. This gets rid of
the current pointless confused effect of dropping rocks instead of
boulders; earth elementals and dust vortices get created instead.

Unlike SliceHack, there is no earthquake for a different beatitude;
instead, varying the beatitude varies the number of monsters created
instead. Also unlike SliceHack, this doesn't really care about
difficulty and won't go out of its way to guarantee certain numbers of
vortices or elementals; the magic of the scroll is what it is.
Furthermore, the generated monsters aren't forced to be hostile.
(Vortices are always hostile anyway; earth elementals might be peaceful
for neutrals, and if so, great.)