Fix: applying hated material tools hurt even wearing gloves
Notably a problem for elvish players applying iron keys. This commit
doesn't make it so that they are immune to keys specifically; what it
does is prevent applying items from causing material damage as long as
the player is wearing gloves, but only so long as the object isn't an
artifact (which should still blast the player, evade grasp, etc.)

This isn't a perfect fix, because applying an artifact made out of a
hated material will still cause material damage even though it should
only trigger artifact effects. Might be noticeable with the Master Key
of Thievery. I didn't see a better way to do it other than a full
refactor of the retouch_object routine to accept more arguments, which I
am unthrilled enough about to not want to do it right now.