Add the Epitaph Patch, allowing players to write their own epitaph
Patch is credited to "RPresser" and comes via SpliceHack; however, the
implementation there contains an information leak (the player learns
that they did indeed create a bones file, which was previously totally
invisible).

To prevent this, I refactored can_make_bones. It is now a totally
deterministic function that checks the bones eligibility of a level, but
doesn't determine whether the bones will actually get generated; the
depth logic that does that is moved into really_done. As a side effect,
wizard mode players are now given the choice to create bones on any
eligible level, bypassing depth considerations (but they are still not
allowed to create bones on straight up ineligible levels, like the
Castle.

For the typical player, this means that you will get prompted to make
your own epitaph whenever you *might* leave bones - not on levels 1-3,
not in the Castle, etc, but everywhere else. However, the epitaph prompt
doesn't necessarily mean that bones will be created.