Fix: weaponless-breaking attack confirmation didn't always trigger
Report was that engulfers could be hit with a weapon, thus breaking the
conduct, with no prompt for confirmation. Since attack_checks had not
only that case, but several others where it would go straight to
attacking without asking for confirmation, the fix is to put the
confirmation logic in a separate function so that it can be called from
those several places. Note that attacking with Stormbringer is the sole
remaining case that will not prompt for confirmation.

Also fixes a related bug in which Stormbringer would trigger a
confirmation for attacking regular hostile monsters, which it shouldn't
do.