Add the Colored Walls Patch
Patch is from SpliceHack but ultimately comes from L back in the 3.4.3
era. The main things include:
 - The Wizard's Tower has magenta walls.
 - Sokoban has blue walls (not in original patch).
 - Vlad's Tower has black walls (not in original patch).
 - The Mines have brown walls (unlike the original patch, all walls are
   universally brown; there is no exception for rooms).
 - Gehennom has orange walls (instead of red, because it has lava).
 - The Astral Plane has white walls.
 - The Valley of the Dead has a rather stronger version: *all* colors
   are suppressed there, becoming shades of gray. Black stuff remains
   black and white stuff remains white, but everything else is
   monochrome. Monsters, objects, etc.

This does not add the colored floors portion of the patch, since it
interacts rather weirdly with dark_room.

This does not currently add the special rooms portion of the patch (as
the original patch does for beehives); since the rtype of a special room
is cleared once you enter it, and this would cause weird behavior like
only the walls of a special room that you saw before entering it would
be colored. It would require a in_rooms counterpart that works with
orig_rtype instead of rtype, which I may or may not add.