Lift restriction on pets fighting grudged monsters of too high level
This is really an update to the Grudge Patch in general. Under previous
rules, a pet would respect its level checks and not attack a monster it
grudged, whereas the monster had no such restriction and would attack
the pet with impunity, until either the hero got it away or the pet
died.

Make it so that if a pet grudges a monster, it will not balk at it being
too high of a level.

Also fix a case in which a pet at less than 25% health would not attack
a monster it grudges, which has the same problem that the grudged
monster will continue to attack it freely.