Augment anti-weaponless-breaking with anti-pacifist-breaking
In the same vein as the previous check, this tries to provide an extra
layer of protection versus breaking pacifist conduct. Attempting to
attack a monster in melee after 100 turns have passed and pacifist is
still intact will prompt for confirmation, unless the player has pressed
F to force-fight.

Only applies to melee, like the anti-weaponless-breaking check.

Change the name of attack_check_weaponless to attack_check_conducts on
account of pacifist being added to it.