Allow any monster eating corpses to possibly get intrinsics
This only applies to gelatinous cubes at the moment, so it probably
doesn't affect much, but it's good to have if any other corpse eating
monsters ever get added.

The fact that this isn't bundled in with the other monster intrinsic
gain serves me right for going through SpliceHack features one by one
without looking for future updates to them.